Commit updated, building bindings
