+++ cmake_bindings/bdsg.sources
@@ -1,12 +1,12 @@
 bdsg.cpp
 std/postypes.cpp
-handlegraph/types.cpp
 std/locale_classes.cpp
 std/ostream_tcc.cpp
+handlegraph/types.cpp
 handlegraph/handle_graph.cpp
 bdsg/is_single_stranded.cpp
 handlegraph/path_handle_graph.cpp
-handlegraph/deletable_handle_graph.cpp
+handlegraph/mutable_handle_graph.cpp
 bdsg/wang_hash.cpp
 bdsg/packed_structs.cpp
 bdsg/utility.cpp
@@ -19,4 +19,3 @@ handlegraph/path_position_handle_graph.cpp
 bdsg/packed_path_position_overlays.cpp
 bdsg/path_position_overlays.cpp
 bdsg/vectorizable_overlays.cpp
-bdsg/vectorizable_overlays_1.cpp
